+++ primefaces-integration-tests/src/test/java/org/primefaces/integrationtests/selectonemenu/SelectOneMenu013Test.java
@@ -1,0 +1,80 @@
+/*
+ * The MIT License
+ *
+ * Copyright (c) 2009-2023 PrimeTek Informatics
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in
+ * all copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+ * THE SOFTWARE.
+ */
+package org.primefaces.integrationtests.selectonemenu;
+
+import java.util.List;
+
+import org.json.JSONObject;
+import org.junit.jupiter.api.Assertions;
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Order;
+import org.junit.jupiter.api.Test;
+import org.openqa.selenium.By;
+import org.openqa.selenium.WebElement;
+import org.openqa.selenium.support.FindBy;
+import org.primefaces.selenium.AbstractPrimePage;
+import org.primefaces.selenium.AbstractPrimePageTest;
+import org.primefaces.selenium.component.SelectOneMenu;
+
+public class SelectOneMenu013Test extends AbstractPrimePageTest {
+
+    @Test
+    @Order(1)
+    @DisplayName("SelectOneMenu: #9677 disabled using advanced table display")
+    public void testDisabled(Page page) {
+        // Arrange
+        SelectOneMenu selectOneMenu = page.selectOneMenu;
+
+        // Act
+        selectOneMenu.show();
+
+        // Assert
+        WebElement table = selectOneMenu.getPanel().findElement(By.className("ui-selectonemenu-table"));
+        Assertions.assertNotNull(table);
+        List<WebElement> options = table.findElements(By.className("ui-selectonemenu-item"));
+
+        assertCss(options.get(0), "ui-selectonemenu-item", "ui-selectonemenu-row", "ui-state-highlight");
+        assertCss(options.get(1), "ui-selectonemenu-item", "ui-selectonemenu-row", "ui-state-disabled");
+        assertCss(options.get(2), "ui-selectonemenu-item", "ui-selectonemenu-row");
+
+        assertConfiguration(selectOneMenu.getWidgetConfiguration());
+    }
+
+    private void assertConfiguration(JSONObject cfg) {
+        assertNoJavascriptErrors();
+        System.out.println("SelectOneMenu Config = " + cfg);
+        Assertions.assertTrue(cfg.has("appendTo"));
+    }
+
+    public static class Page extends AbstractPrimePage {
+        @FindBy(id = "form:selectonemenu")
+        SelectOneMenu selectOneMenu;
+
+        @Override
+        public String getLocation() {
+            return "selectonemenu/selectOneMenu013.xhtml";
+        }
+    }
+
+}