+++ primefaces/src/main/java/org/primefaces/validate/bean/PositiveOrZeroClientValidationConstraint.java
@@ -31,8 +31,8 @@ public class PositiveOrZeroClientValidationConstraint extends AbstractClientVali
 
     public static final String CONSTRAINT_ID = "PositiveOrZero";
     public static final String MESSAGE_METADATA = "data-p-positiveorzero-msg";
-    public static final String CONSTRAINT_CLASS_NAME = CONSTRAINT_PACKAGE  + "." + CONSTRAINT_ID;
-    public static final String MESSAGE_ID = "{" + CONSTRAINT_CLASS_NAME + ".message}";
+    public static final String CONSTRAINT_CLASS_NAME = String.format("%s.%s", javax.validation.constraints.NotNull.class.getPackage().getName(), CONSTRAINT_ID);
+    public static final String MESSAGE_ID =  String.format("{%s.%s}", CONSTRAINT_CLASS_NAME, ATTR_MESSAGE);
     public static final String MIN_VALUE = "0";
 
     public PositiveOrZeroClientValidationConstraint() {