+++ primefaces/src/main/resources/META-INF/resources/primefaces/toggleswitch/toggleswitch.js
@@ -54,14 +54,12 @@ PrimeFaces.widget.ToggleSwitch = PrimeFaces.widget.BaseWidget.extend({
             $this.jq.removeClass('ui-toggleswitch-focus');
         })
         .on('keydown.toggleSwitch', function(e) {
-            var keyCode = $.ui.keyCode;
-            if(e.which === keyCode.SPACE) {
+            if(e.key === ' ') {
                 e.preventDefault();
             }
         })
         .on('keyup.toggleSwitch', function(e) {
-            var keyCode = $.ui.keyCode;
-            if(e.which === keyCode.SPACE) {
+            if(e.key === ' ') {
                 $this.toggle();
 
                 e.preventDefault();